WIP select parent field when last child is deleted
smartcitiesdata/smartcitiesdata#653

+++ apps/andi/lib/andi/input_schemas/data_dictionary_fields.ex
@@ -24,14 +24,16 @@ defmodule Andi.InputSchemas.DataDictionaryFields do
   end
 
   def remove_field(existing_field_id) do
-    existing_field = Repo.get!(DataDictionary, existing_field_id)
+    existing_field = Repo.get(DataDictionary, existing_field_id)
 
-    case Repo.delete(existing_field) do
-      {:error, _changeset} ->
-        {:error, DataDictionary.changeset(%DataDictionary{}, existing_field)}
+    if existing_field != nil do
+      case Repo.delete(existing_field) do
+        {:error, _changeset} ->
+          {:error, DataDictionary.changeset(%DataDictionary{}, existing_field)}
 
-      good ->
-        good
+        good ->
+          good
+      end
     end
   end
